torch_dtype change from float16 to bfloat16

+++ docker/qwen_processor.py
@@ -14,7 +14,7 @@ class QwenProcessor:
             MODEL_NAME,
             device_map="auto",
             trust_remote_code=True,
-            torch_dtype=torch.float16
+            torch_dtype=torch.bfloat16  # ← FIXED: Changed from float16 to bfloat16
         )
         
         self.processor = AutoProcessor.from_pretrained(
